cleanup: refactored scattered scripts

+++ ssl/reset.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 
-# Reset
+# confirmation dialog
+read -p "Are you sure you want to reset the CA? [y/N] " -n 1 -r
+echo   # (optional) move to a new line
+if [[ ! $REPLY =~ ^[Yy]$ ]]
+then
+    exit 1
+fi
 
 # rootCA
 rm rootCA -rf
